suppress warning on Solaris
git-svn-id: https://svn.r-project.org/R/trunk@33057 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/gram.c
@@ -3662,7 +3662,7 @@ int isValidName(char *name)
         if (wc == L'.')
         {
             /* We don't care about other than ASCII digits */
-            if (isdigit(*p))
+            if (isdigit((int)*p))
                 return 0;
             /* Mbrtowc(&wc, p, n, NULL); if(iswdigit(wc)) return 0; */
         }
